+++ zcord/models/message.py
@@ -9,6 +9,7 @@ from zcord.missing import MISSING
 from zcord.models.attachment import Attachment
 from zcord.models.channel import Channel
 from zcord.models.embed import Embed
+from zcord.models.reaction import Reaction
 from zcord.models.snowflake import Snowflake
 from zcord.models.user import User
 from zcord.utils import from_payload
@@ -109,7 +110,7 @@ class Message:
     type: MessageType
     edited_timestamp: datetime | None = None
     mention_channels: list[Channel] | MISSING = MISSING
-    reactions: list | MISSING = MISSING
+    reactions: list[Reaction] | MISSING = MISSING
     webhook_id: Snowflake | MISSING = MISSING
     activity: Any | MISSING = MISSING
     application: Any | MISSING = MISSING
@@ -144,6 +145,7 @@ class Message:
             embeds=Embed._from_payload,
             mention_roles=Snowflake,
             mention_channels=Channel._from_payload,
+            reactions=Reaction._from_payload,
             webhook_id=Snowflake,
             type=MessageType,
             application_id=Snowflake,